Extract step FX logic into dedicated module
